Refatorando migrations

+++ database/migrations/2024_05_13_231936_create_servicos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('servicos', function (Blueprint $table) {
             $table->id();
             $table->string('descricao');
-            $table->float('valor');
+            $table->integer('valor');
             $table->timestamps();
         });
     }
